requirements.txt
requests
django
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
cloudinary